pyproject.toml
[project]
name = "inventory"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.11",
    "jinja2>=3.1.6",
    "pydantic-settings>=2.8.1",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.24",
]

[tool.ruff]
line-length = 100
lint.select = ["ALL"]
lint.ignore = ["D", "FAST003"]

[tool.pyright]
typeCheckingMode = "strict"
